pyproject.toml
[tool.poetry]
name = "agent-template"
version = "0.1.0"
description = "A template for creating FastAPI-based agents"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
# Core dependencies - automatically installed by bin/setup
python = "^3.11"  # deepagents requires Python >=3.11
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
deepagents = "^0.3.6"  # Latest version with create_deep_agent
langgraph = "^0.2.0"  # Required by deepagents
python-dotenv = "^1.0.0"
pymongo = "^4.6.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
certifi = "^2023.11.17"
openai = "^1.0.0"
redis = {extras = ["asyncio"], version = "^5.0.0"}
fastapi-cache2 = "^0.2.1"

[tool.poetry.group.dev.dependencies]
# Dev dependencies - automatically installed by bin/setup
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
httpx = "^0.25.0"
black = "^23.12.0"
mypy = "^1.8.0"
ruff = "^0.1.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.ruff]
line-length = 100
target-version = "py311"
